initial selector file renamed
